fix: collapse editorial panels by default
